test(agent): keep worker failures at owning boundaries
