+++ packages/junior/tests/integration/slack/conversation-turn-steering-behavior.test.ts
@@ -787,51 +787,4 @@ describe("Slack behavior: durable turn steering", () => {
       ]),
     );
   });
-
-  it("keeps the mailbox pending when the agent fails before input commit", async () => {
-    const state = getStateAdapter();
-    const executeAgentRun: AgentRunner["run"] = async (request) => {
-      expect(request.durability?.onInputCommitted).toEqual(
-        expect.any(Function),
-      );
-      throw new Error("agent crashed before input commit");
-    };
-    const { conversationId, queue, runNextQueuedWork, services } =
-      createTurnHarness({
-        agentRunner: { run: executeAgentRun },
-        state,
-      });
-
-    await expect(
-      handleSlackWebhookAndFlush({
-        request: slackWebhookRequest(
-          makeMessageEvent({
-            eventType: "app_mention",
-            text: `<@${SLACK_BOT_USER_ID}> start the incident summary`,
-            ts: THREAD_TS,
-          }),
-        ),
-        services,
-      }),
-    ).resolves.toMatchObject({ status: 200 });
-
-    await expect(runNextQueuedWork()).resolves.toEqual({
-      status: "pending_requeued",
-    });
-
-    const work = await getConversationWorkState({
-      conversationId,
-      state,
-    });
-    expect(work?.needsRun).toBe(true);
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(1);
-    expect(work?.messages[0]?.injectedAtMs).toBeUndefined();
-    expect(queue.sentRecords()).toEqual([
-      expect.objectContaining({ conversationId }),
-      expect.objectContaining({
-        conversationId,
-        idempotencyKey: expect.stringContaining(`pending:${conversationId}:`),
-      }),
-    ]);
-  });
 });
